requirements.txt
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
boto3==1.9.144
botocore==1.12.144
certifi==2019.3.9
decorator==4.3.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.8
django-allauth==0.39.1
django-disqus==0.5
django-extensions==2.2.1
django-filter==2.2.0
django-storages==1.7.1
djangorestframework==3.10.3
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.14
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
mysql-connector-python==8.0.15
mysqlclient==1.4.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
oauthlib==3.0.1
openpyxl==2.6.1
pandocfilters==1.4.2
parso==0.3.4
pbr==5.2.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2-binary==2.8.2
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
typed-ast==1.3.2
urllib3==1.24.3
virtualenv==16.5.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.3
widgetsnbextension==3.4.2
wrapt==1.11.1
XlsxWriter==1.1.7